[data/utils.py]create utils.py
